diktatFix + review notes
