+++ save-backend/src/main/kotlin/com/saveourtool/save/backend/controllers/FileController.kt
@@ -3,7 +3,7 @@ package com.saveourtool.save.backend.controllers
 import com.saveourtool.save.backend.ByteBufferFluxResponse
 import com.saveourtool.save.backend.StringResponse
 import com.saveourtool.save.backend.service.ProjectService
-import com.saveourtool.save.backend.storage.FileStorage
+import com.saveourtool.save.backend.storage.MigrationFileStorage
 import com.saveourtool.save.configs.ApiSwaggerSupport
 import com.saveourtool.save.domain.*
 import com.saveourtool.save.entities.Project
@@ -37,7 +37,7 @@ import reactor.core.publisher.Mono
     Tag(name = "files"),
 )
 class FileController(
-    private val fileStorage: FileStorage,
+    private val fileStorage: MigrationFileStorage,
     private val projectService: ProjectService,
 ) {
     @Operation(
