+++ save-backend/src/main/kotlin/com/saveourtool/save/backend/controllers/internal/FileInternalController.kt
@@ -2,7 +2,7 @@ package com.saveourtool.save.backend.controllers.internal
 
 import com.saveourtool.save.backend.ByteBufferFluxResponse
 import com.saveourtool.save.backend.controllers.DownloadFilesController
-import com.saveourtool.save.backend.storage.FileStorage
+import com.saveourtool.save.backend.storage.MigrationFileStorage
 import com.saveourtool.save.domain.FileKey
 import com.saveourtool.save.domain.ProjectCoordinates
 import com.saveourtool.save.utils.switchIfEmptyToNotFound
@@ -17,14 +17,14 @@ import reactor.core.publisher.Mono
 import reactor.kotlin.core.publisher.toMono
 
 /**
- * Internal controller for [FileStorage]
+ * Internal controller for [MigrationFileStorage]
  *
  * @property fileStorage
  */
 @RestController
 @RequestMapping("/internal/files")
 class FileInternalController(
-    private val fileStorage: FileStorage,
+    private val fileStorage: MigrationFileStorage,
 ) {
     /**
      * @param organizationName
